ns: move IFactoryResetInterface
